Please remove our public resolver from the list
Hello,

>DNS resolver of the associative French ISP FDN accessible over DoH.
We aren't the French ISP FDN at all. We are hosting an instance of `doh-proxy` (by jedisct1) that redirects DoH queries we're receiving to FDN's open DNS resolver, but we are a fully separate entity.

> Proxies to Google DNS.
We've never proxied anything to Google DNS.

More information here: https://42l.fr/Service-DoH
(you might not be able to access the page if you're using dnscrypt-proxy)

We've been suffering from an *heavy* traffic from dnscrypt-proxy users. Our little infrastructure isn't ready to have its place on this list.

We'd be glad if you can remove our proxy from it.

As a proof we're not impersonating the association, we sent an encrypted email signed with our PGP key to jedisct1's pureftpd address (sorry, was it the wrong one?).
You can check our PGP key here : https://42l.fr/Contact

Thanks for your understanding.