plan9dns-nj lives on (#804)
ok, was traveling to Guatemala and everything was shut down for Good Friday, finally got internet access and saw that the nj server was locked up drawing 100% cpu for about 5 days! Rebooted and looks good. Vultr has been doing alot of network upgrades and sometimes just halts the server and then powers it back on.... Sorry for the extra work.